Added small changes to the data types of the Shipping, ShippingDTO,...

See merge request mariusrd/elektromart-ecom-website!42

+++ src/main/java/com/electromart/electromart/repository/ShippingRepository.java
@@ -3,5 +3,5 @@ package com.electromart.electromart.repository;
 import com.electromart.electromart.entity.Shipping;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShippingRepository extends JpaRepository<Shipping, Integer> {
+public interface ShippingRepository extends JpaRepository<Shipping, Long> {
 }
